refactor: move plugin protocols into protocols.py
Extract the runtime-checkable plugin protocols, the `BuildState` literal,
and the `BUILD_STATES` set out of `loader.py` into a new leaf module
`protocols.py`. `loader.py` and `__main__.py` import from it; docs and
tests point at the new canonical location.
