+++ src/dynamic_metadata/loader.py
@@ -12,11 +12,7 @@ from types import MappingProxyType
 from typing import (
     TYPE_CHECKING,
     Any,
-    Literal,
-    Protocol,
     cast,
-    get_args,
-    runtime_checkable,
 )
 
 from ._compat import metadata
@@ -28,11 +24,12 @@ from .info import (
     LIST_STR_FIELDS,
     SCALAR_FIELDS,
 )
-
-BuildState = Literal[
-    "sdist", "wheel", "editable", "metadata_wheel", "metadata_editable"
-]
-BUILD_STATES: frozenset[str] = frozenset(get_args(BuildState))
+from .protocols import (
+    BUILD_STATES,
+    BuildState,
+    DynamicMetadataBuildStateProtocol,
+    DynamicMetadataProtocol,
+)
 
 if TYPE_CHECKING:
     from collections.abc import Generator, Sequence
@@ -42,7 +39,6 @@ if TYPE_CHECKING:
 
 
 __all__ = [
-    "BuildState",
     "load_dynamic_metadata",
     "load_provider",
     "process_dynamic_metadata",
@@ -51,32 +47,6 @@ __all__ = [
 
 def __dir__() -> list[str]:
     return __all__
-
-
-@runtime_checkable
-class DynamicMetadataProtocol(Protocol):
-    def dynamic_metadata(
-        self,
-        settings: Mapping[str, Any],
-        project: Mapping[str, Any],
-    ) -> dict[str, Any]: ...
-
-
-@runtime_checkable
-class DynamicMetadataBuildStateProtocol(DynamicMetadataProtocol, Protocol):
-    def build_state(self, build_state: BuildState) -> None: ...
-
-
-@runtime_checkable
-class DynamicMetadataRequirementsProtocol(DynamicMetadataProtocol, Protocol):
-    def get_requires_for_dynamic_metadata(
-        self, settings: Mapping[str, Any]
-    ) -> list[str]: ...
-
-
-@runtime_checkable
-class DynamicMetadataWheelProtocol(DynamicMetadataProtocol, Protocol):
-    def dynamic_wheel(self, settings: Mapping[str, Any]) -> dict[str, bool]: ...
 
 
 # Entry-point group a plugin distribution registers a named provider under. The
